+++ OsmAnd-java/src/main/java/net/osmand/binary/CachedOsmandIndexes.java
@@ -246,14 +246,14 @@ public class CachedOsmandIndexes {
 
 		for (MapPart index : found.getMapIndexList()) {
 			MapIndex mi = new MapIndex();
-			mi.length = (int) index.getSize();
-			mi.filePointer = (int) index.getOffset();
+			mi.length = index.getSize();
+			mi.filePointer = index.getOffset();
 			mi.name = index.getName();
 
 			for (MapLevel mr : index.getLevelsList()) {
 				MapRoot root = new MapRoot();
-				root.length = (int) mr.getSize();
-				root.filePointer = (int) mr.getOffset();
+				root.length = mr.getSize();
+				root.filePointer = mr.getOffset();
 				root.left = mr.getLeft();
 				root.right = mr.getRight();
 				root.top = mr.getTop();
@@ -269,15 +269,15 @@ public class CachedOsmandIndexes {
 
 		for (AddressPart index : found.getAddressIndexList()) {
 			AddressRegion mi = new AddressRegion();
-			mi.length = (int) index.getSize();
-			mi.filePointer = (int) index.getOffset();
+			mi.length = index.getSize();
+			mi.filePointer = index.getOffset();
 			mi.name = index.getName();
 			mi.enName = index.getNameEn();
 			mi.indexNameOffset = index.getIndexNameOffset();
 			for (CityBlock mr : index.getCitiesList()) {
 				CitiesBlock cblock = new CitiesBlock();
-				cblock.length = (int) mr.getSize();
-				cblock.filePointer = (int) mr.getOffset();
+				cblock.length = mr.getSize();
+				cblock.filePointer = mr.getOffset();
 				cblock.type = mr.getType();
 				mi.cities.add(cblock);
 			}
@@ -288,8 +288,8 @@ public class CachedOsmandIndexes {
 
 		for (PoiPart index : found.getPoiIndexList()) {
 			PoiRegion mi = new PoiRegion();
-			mi.length = (int) index.getSize();
-			mi.filePointer = (int) index.getOffset();
+			mi.length = index.getSize();
+			mi.filePointer = index.getOffset();
 			mi.name = index.getName();
 			mi.left31 = index.getLeft();
 			mi.right31 = index.getRight();
@@ -301,8 +301,8 @@ public class CachedOsmandIndexes {
 
 		for (TransportPart index : found.getTransportIndexList()) {
 			TransportIndex mi = new TransportIndex();
-			mi.length = (int) index.getSize();
-			mi.filePointer = (int) index.getOffset();
+			mi.length = index.getSize();
+			mi.filePointer = index.getOffset();
 			mi.name = index.getName();
 			mi.left = index.getLeft();
 			mi.right = index.getRight();
@@ -321,14 +321,14 @@ public class CachedOsmandIndexes {
 
 		for (RoutingPart index : found.getRoutingIndexList()) {
 			RouteRegion mi = new RouteRegion();
-			mi.length = (int) index.getSize();
-			mi.filePointer = (int) index.getOffset();
+			mi.length = index.getSize();
+			mi.filePointer = index.getOffset();
 			mi.name = index.getName();
 
 			for (RoutingSubregion mr : index.getSubregionsList()) {
 				RouteSubregion sub = new RouteSubregion(mi);
-				sub.length = (int) mr.getSize();
-				sub.filePointer = (int) mr.getOffset();
+				sub.length = mr.getSize();
+				sub.filePointer = mr.getOffset();
 				sub.left = mr.getLeft();
 				sub.right = mr.getRight();
 				sub.top = mr.getTop();
@@ -346,8 +346,8 @@ public class CachedOsmandIndexes {
 		
 		for (HHRoutingPart index : found.getHhRoutingIndexList()) {
 			HHRouteRegion mi = new HHRouteRegion();
-			mi.length = (int) index.getSize();
-			mi.filePointer = (int) index.getOffset();
+			mi.length = index.getSize();
+			mi.filePointer = index.getOffset();
 			mi.edition = index.getEdition();
 			mi.profile = index.getProfile();
 			mi.profileParams = index.getProfileParamsList();
